feat: FE 배포 후 deprecated API 제거
- GalleryWriteEndpoint 삭제하고 GalleryDeleteEndpoint로 분리
- 갤러리 관련 API 구조 정리
- deprecated된 필드 및 메서드 제거
- API 응답 DTO 정리
- TokenQueryDto 생성자 호출 수정
- Checkstyle 경고 수정
