+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/auth/service/LoginService.java
@@ -33,9 +33,7 @@ public class LoginService {
 
     var tokenQueryDTO = new TokenQueryDto(
         tokens.getAccessToken(),
-        tokens.getAccessTokenExpireAt(),
         tokens.getRefreshToken(),
-        tokens.getRefreshTokenExpireAt(),
         tokens.getSocialToken()
     );
