+++ services/lifepuzzle-api/src/main/java/io/itmca/lifepuzzle/domain/content/endpoint/GalleryQueryEndpoint.java
@@ -17,14 +17,6 @@ import org.springframework.web.bind.annotation.RestController;
 public class GalleryQueryEndpoint {
   private final GalleryQueryService galleryQueryService;
 
-  @Deprecated
-  @Operation(summary = "홈 화면 조회")
-  @GetMapping({"/v1/heroes/{heroId}/gallery"})
-  public ResponseEntity<GalleryQueryResponse> getHeroGallery(@PathVariable("heroId") Long heroId) {
-    var response = galleryQueryService.getHeroGallery(heroId);
-    return ResponseEntity.ok(response);
-  }
-
   @Operation(summary = "홈 화면 조회")
   @GetMapping({"/v1/galleries"})
   public ResponseEntity<GalleryQueryResponse> getHeroGalleryV2(@RequestParam("heroId") Long heroId) {
